Fix overrun in README
The README includes an example that allocates a single byte
buffer an sprintf's to that. This might work due to a quirk
with aligning to 8-byte boundaries but "%02x " is always
going to be 2 characters + 2 hex + 1 space + nul (7 characters)
which is a little dicey.

snprintf makes this safer because snprintf accounts for the
buffer length.